fix #7 bug in my first git clone
